+++ tests/test_grouped_env/expected_result_i_placement.csv
